Work In Progress: Fixing Anchor Tags
In realizing the anchor tags within the unorder list are not working, I have begun an investigation as why this is. One solution, would be to remove the ul and just use anchor tags, but this is not the preferred method at the time as this would require further refactoring of styles.